extract select button
